+++ lib/features/receipts/presentation/receipt_route.dart
@@ -4,10 +4,11 @@ import 'package:provider/provider.dart';
 
 import '../application/provider_receipts.dart';
 import '../domain/receipt.dart';
-import 'show_jpg_dialog.dart';
+import 'show_image_dialog.dart';
 //import '../../../utils/show_pdf_dialog.dart';
 import '../../snacks/snacks.dart';
 import '../../../widgets/data_widget.dart';
+import '../../../widgets/confirm_dialog.dart';
 
 /// Route for showing a single Receipt
 ///
@@ -45,13 +46,12 @@ class _ReceiptRouteState extends State<ReceiptRoute> {
     try {
       final file = await providerReceipts.getReceiptImage(widget.receipt);
       switch (file.runtimeType) {
-        // FIXME Figure out what this means
-        case Image:
+        case == Image:
           if (context.mounted) {
             showDialog(
               context: context,
               builder: (BuildContext context) {
-                return ShowJpgDialog(
+                return ShowImageDialog(
                   title: widget.receipt.fileName,
                   image: file,
                 );
@@ -75,7 +75,6 @@ class _ReceiptRouteState extends State<ReceiptRoute> {
           break;
       }
     } catch (e) {
-      // TODO Show error snackbar
       debugPrint(e.toString());
     }
   }
@@ -134,8 +133,7 @@ class _ReceiptRouteState extends State<ReceiptRoute> {
               _description(
                   title: locals.description,
                   content: widget.receipt.description),
-              // FIXME Localization
-              DataWidget(title: 'Minute', content: widget.receipt.minute),
+              DataWidget(title: locals.minute, content: widget.receipt.minute),
               Text(
                 locals.file,
                 style: Theme.of(context).textTheme.displaySmall,
@@ -151,7 +149,19 @@ class _ReceiptRouteState extends State<ReceiptRoute> {
                 ),
               ),
               ElevatedButton(
-                onPressed: () {
+                onPressed: () async {
+                  final confirmation = await showDialog<bool>(
+                      context: context,
+                      builder: (BuildContext context) {
+                        return ConfirmDialog(
+                          child: Text(
+                            locals.receiptRouteConfirmDeleteText,
+                            style: Theme.of(context).textTheme.titleLarge,
+                          ),
+                        );
+                      });
+                  if (confirmation == null) return;
+                  if (!confirmation) return;
                   providerReceipts.deleteReceipt(widget.receipt).then((value) {
                     Navigator.of(context).pop();
                     sendSnack(
